Gracefully handle list that don't end with nil
Using dotted pair notation, elisp lets you create lists that don't end in nil.
Regular lists usually look like this:

'(rose violet buttercup)

         --- ---      --- ---      --- ---
        |   |   |--> |   |   |--> |   |   |--> nil
         --- ---      --- ---      --- ---
          |            |            |
          |            |            |
           --> rose     --> violet   --> buttercup

But you can make them end on a 'non sequence' object, instead of a last `cons`
with nil ending.  For example:

(rose violet . buttercup)

         --- ---      --- ---
        |   |   |--> |   |   |--> buttercup
         --- ---      --- ---
          |            |
          |            |
           --> rose     --> violet

This case made the previous `treepy-right` implementation to throw an exception,
given that it uses `seq-let` to deconstruct lists.

When `seq-let` faces an object that doesn't satisfy `sequencep` (like
`'buttercup` in this example) it fails to destructure the list and throws a
`(wrong-type-argument sequencep butterfly)` error.

In order to prevent this, we check that the thing we're about to deconstruct is
actually a sequence, and if not, we `cons` it to nil, so that `seq-let` is able
to process it.