Use `defun' indentation type for treepy--context-assoc
This helps avoiding long lines and looks better even in cases were the
line length isn't a problem.  An alternative that could be used to only
deal with the long-line cases would be to place the first argument on a
separate line, but that would waste vertical space, look odd, and would
needlessly cause different uses of this function to look different.